requirements.txt
PyMuPDF
PyPDF2
aws-opentelemetry-distro>=0.14.0
aws_requests_auth
aws_sdk_bedrock_runtime
awscli>=1.42.14
bedrock-agentcore-starter-toolkit>=0.2.3
bedrock-agentcore>=1.1.1
boto3>=1.40.8
botocore>=1.40.8
cfn-flip
ddgs
dill>=0.4.0,<0.5.0
fpdf
iprogress
ipykernel
ipywidgets
itables
matplotlib
mcp
nbformat
numpy==2.2.4
opensearch-py
pandas
pillow
pypdf
pyyaml
requests
requests-aws4auth
retrying
rx==3.2.0
scipy
seaborn
strands-agents-tools==0.2.7
strands-agents==1.8.0
streamlit
streamlit-cognito-auth
termcolor
uv
websockets==15.0.1
wheel
